+++ contracts/Token.sol
@@ -30,7 +30,8 @@ contract Token is ERC2771Context, ERC20Permit, Pausable, AccessControl {
     /// @notice The role that has the ability to burn tokens from anywhere. Usage is planned for legal purposes and error recovery.
     bytes32 public constant BURNER_ROLE = keccak256("BURNER_ROLE");
     /// @notice The role that has the ability to grant transfer rights to other addresses
-    bytes32 public constant TRANSFERERADMIN_ROLE = keccak256("TRANSFERERADMIN_ROLE");
+    bytes32 public constant TRANSFERERADMIN_ROLE =
+        keccak256("TRANSFERERADMIN_ROLE");
     /// @notice Addresses with this role do not need to satisfy any requirements to send or receive tokens
     bytes32 public constant TRANSFERER_ROLE = keccak256("TRANSFERER_ROLE");
     /// @notice The role that has the ability to pause the token
@@ -67,7 +68,7 @@ contract Token is ERC2771Context, ERC20Permit, Pausable, AccessControl {
     /// @notice defines the maximum amount of tokens that can be minted by a specific minter. If zero, no tokens can be minted.
     mapping(address => uint256) public mintingAllowance; // used for token generating events such as vesting or new financing rounds
 
-    event RequirementsChanged(uint newRequirements);
+    event RequirementsChanged(uint256 newRequirements);
     event AllowListChanged(AllowList indexed allowList);
     event MintingAllowanceChanged(address indexed minter, uint256 newAllowance);
 
@@ -80,7 +81,18 @@ contract Token is ERC2771Context, ERC20Permit, Pausable, AccessControl {
     @param _requirements requirements an address has to meet for sending or receiving tokens
     @param _admin address of the admin. Admin will initially have all roles and can grant roles to other addresses.
     */
-    constructor(address _trustedForwarder, address _admin, AllowList _allowList, uint256 _requirements, string memory _name, string memory _symbol) ERC2771Context(_trustedForwarder) ERC20Permit(_name) ERC20(_name, _symbol) {
+    constructor(
+        address _trustedForwarder,
+        address _admin,
+        AllowList _allowList,
+        uint256 _requirements,
+        string memory _name,
+        string memory _symbol
+    )
+        ERC2771Context(_trustedForwarder)
+        ERC20Permit(_name)
+        ERC20(_name, _symbol)
+    {
         // Grant admin roles
         _setupRole(DEFAULT_ADMIN_ROLE, _admin); // except for the Minter and Transferer role, the _admin is the roles admin for all other roles
         _setRoleAdmin(MINTER_ROLE, MINTERADMIN_ROLE);
@@ -97,12 +109,18 @@ contract Token is ERC2771Context, ERC20Permit, Pausable, AccessControl {
         requirements = _requirements;
     }
 
-    function setAllowList(AllowList _allowList) public onlyRole(DEFAULT_ADMIN_ROLE) {
+    function setAllowList(AllowList _allowList)
+        public
+        onlyRole(DEFAULT_ADMIN_ROLE)
+    {
         allowList = _allowList;
         emit AllowListChanged(_allowList);
     }
 
-    function setRequirements(uint256 _requirements) public onlyRole(REQUIREMENT_ROLE) {
+    function setRequirements(uint256 _requirements)
+        public
+        onlyRole(REQUIREMENT_ROLE)
+    {
         requirements = _requirements;
         emit RequirementsChanged(_requirements);
     }
@@ -117,15 +135,25 @@ contract Token is ERC2771Context, ERC20Permit, Pausable, AccessControl {
         @param _minter address of the minter contract
         @param _allowance maximum amount of tokens that can be minted by this minter IN THIS ROUND
     */
-    function setUpMinter(address _minter, uint256 _allowance) public onlyRole(getRoleAdmin(MINTER_ROLE)){
+    function setUpMinter(address _minter, uint256 _allowance)
+        public
+        onlyRole(getRoleAdmin(MINTER_ROLE))
+    {
         _grantRole(MINTER_ROLE, _minter);
         require(mintingAllowance[_minter] == 0 || _allowance == 0); // to prevent frontrunning when setting a new allowance, see https://www.adrianhetman.com/unboxing-erc20-approve-issues/
         mintingAllowance[_minter] = _allowance;
         emit MintingAllowanceChanged(_minter, _allowance);
     }
 
-    function mint(address _to, uint256 _amount) public onlyRole(MINTER_ROLE) returns (bool) {
-        require(mintingAllowance[_msgSender()] >= _amount, "MintingAllowance too low");
+    function mint(address _to, uint256 _amount)
+        public
+        onlyRole(MINTER_ROLE)
+        returns (bool)
+    {
+        require(
+            mintingAllowance[_msgSender()] >= _amount,
+            "MintingAllowance too low"
+        );
         mintingAllowance[_msgSender()] -= _amount;
         _mint(_to, _amount);
         return true;
@@ -147,11 +175,16 @@ contract Token is ERC2771Context, ERC20Permit, Pausable, AccessControl {
         super._beforeTokenTransfer(_from, _to, _amount);
         _requireNotPaused();
         require(
-            hasRole(BURNER_ROLE, _msgSender()) || hasRole(TRANSFERER_ROLE, _from) || allowList.map(_from) & requirements == requirements || _from == address(0),
+            hasRole(BURNER_ROLE, _msgSender()) ||
+                hasRole(TRANSFERER_ROLE, _from) ||
+                allowList.map(_from) & requirements == requirements ||
+                _from == address(0),
             "Sender is not allowed to transact. Either locally issue the role as a TRANSFERER or they must meet requirements as defined in the allowList"
         ); // address(0), because this is the _from address in case of minting new tokens
         require(
-            hasRole(TRANSFERER_ROLE, _to) || allowList.map(_to) & requirements == requirements || _to == address(0),
+            hasRole(TRANSFERER_ROLE, _to) ||
+                allowList.map(_to) & requirements == requirements ||
+                _to == address(0),
             "Receiver is not allowed to transact. Either locally issue the role as a TRANSFERER or they must meet requirements as defined in the allowList"
         ); // address(0), because this is the _to address in case of burning tokens
     }
@@ -166,15 +199,25 @@ contract Token is ERC2771Context, ERC20Permit, Pausable, AccessControl {
 
     /**
      * @dev both ERC20Pausable and ERC2771Context have a _msgSender() function, so we need to override and select which one to use.
-     */ 
-    function _msgSender() internal view override(Context, ERC2771Context) returns (address) {
+     */
+    function _msgSender()
+        internal
+        view
+        override(Context, ERC2771Context)
+        returns (address)
+    {
         return ERC2771Context._msgSender();
     }
 
     /**
      * @dev both ERC20Pausable and ERC2771Context have a _msgData() function, so we need to override and select which one to use.
      */
-    function _msgData() internal view override(Context, ERC2771Context) returns (bytes calldata) {
+    function _msgData()
+        internal
+        view
+        override(Context, ERC2771Context)
+        returns (bytes calldata)
+    {
         return ERC2771Context._msgData();
     }
 }
